requirements.txt
streamlit
semantic-kernel
python-dotenv
